+++ packages/paraglide/paraglide-next/example/src/middleware.ts
@@ -1,6 +1,0 @@
-import { NextRequest } from "next/server";
-import * as paraglide from "./paraglide/adapter";
-
-export function middleware(request: NextRequest) {
-	return paraglide.middleware(request);
-}