+++ packages/plugins/i18next/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inlang/plugin-i18next
 
+## 5.0.10
+
+### Patch Changes
+
+- Updated dependencies [c0b857a]
+- Updated dependencies [91ba4eb]
+  - @inlang/sdk@2.3.0
+
 ## 5.0.9
 
 ### Patch Changes